post

+++ app/Http/Controllers/TasksController.php
@@ -15,14 +15,28 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'asc')->paginate(10);
 
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        // Welcomeビューでそれらを表示
+        return view('tasks.index', $data);
     }
+        
+        // タスク一覧を取得
+        //$tasks = Task::orderBy('id', 'desc')->paginate(10);
+        // タスク一覧ビューでそれを表示
+        //   return view('tasks.index', [
+        //  'tasks' => $tasks,
+        //]);
 
     /**
      * Show the form for creating a new resource.
@@ -32,11 +46,11 @@ class TasksController extends Controller
     public function create()
     {
         $task = new Task;
-
         // タスク作成ビューを表示
         return view('tasks.create', [
             'task' => $task,
         ]);
+        
     }
 
     /**
@@ -55,10 +69,11 @@ class TasksController extends Controller
 
         // タスクを作成
         $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
+    
+            $task->user_id = \Auth::id();
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -71,13 +86,14 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        // idの値でタスクを検索して取得
+        // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
 
-        // タスク詳細ビューでそれを表示
+        // メッセージ詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
         ]);
+        
     }
 
     /**
@@ -114,11 +130,12 @@ class TasksController extends Controller
         
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        $task->status = $request->status;
-        // タスクを更新
-        $task->content = $request->content;
-        $task->save();
-
+        
+            $task->user_id = $request->user_id;
+            $task->status = $request->status;
+            // タスクを更新
+            $task->content = $request->content;
+            $task->save();
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -131,12 +148,16 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
 
-        // トップページへリダイレクトさせる
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 }
+
